move exports/main -> lib
