+++ exports/main.ts
@@ -1,5 +1,5 @@
-export * from "./main/intrinsics.ts";
-export * from "./main/json.ts";
-export * from "./main/parameters.ts";
-export * from "./main/policy.ts";
-export * from "./main/template.ts";
+export * from "../lib/intrinsics.ts";
+export * from "../lib/json.ts";
+export * from "../lib/parameters.ts";
+export * from "../lib/policy.ts";
+export * from "../lib/template.ts";
