+++ lib/intrinsics.ts
